# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users/{id}": {
            "get": {
                "description": "get user with current sessions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current urser data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Authentication token",
                        "name": "auth",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/entities.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/entities.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entities.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Event": {
            "type": "object",
            "required": [
                "description",
                "end_date",
                "location",
                "name",
                "public",
                "start_date"
            ],
            "properties": {
                "cashiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.User"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "free_wifi": {
                    "type": "boolean"
                },
                "guests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Guest"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.EventImage"
                    }
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/entities.LocalisedTextContent"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.EventProduct"
                    }
                },
                "public": {
                    "type": "boolean"
                },
                "sellers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.User"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "ticketTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.TicketType"
                    }
                },
                "tickets_amount": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "walletType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.WalletType"
                    }
                }
            }
        },
        "entities.EventImage": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "eventID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entities.EventProduct": {
            "type": "object",
            "required": [
                "name",
                "token_price"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "eventID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "token_price": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.Guest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "guestGroupID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Ticket"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Wallet"
                    }
                }
            }
        },
        "entities.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "entities.Language": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "languageID": {
                    "type": "integer"
                },
                "languageType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.LocalisedTextContent": {
            "type": "object",
            "properties": {
                "contentID": {
                    "type": "integer"
                },
                "contentType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "translations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.LocalisedTextTranslation"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.LocalisedTextTranslation": {
            "type": "object",
            "required": [
                "content",
                "language"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "$ref": "#/definitions/entities.Language"
                },
                "localisedTextContentID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.Ticket": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "guestID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ticketType": {
                    "$ref": "#/definitions/entities.TicketType"
                },
                "ticketTypeID": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.TicketTransaction"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "entities.TicketTransaction": {
            "type": "object",
            "required": [
                "quantity",
                "type"
            ],
            "properties": {
                "cashierID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "ticketID": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.TicketType": {
            "type": "object",
            "required": [
                "end_validity_date",
                "name",
                "start_validity_date"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "end_validity_date": {
                    "type": "string"
                },
                "eventID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "start_validity_date": {
                    "type": "string"
                },
                "unlimited": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.Token": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.TokenTransaction"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "walletID": {
                    "type": "integer"
                }
            }
        },
        "entities.TokenTransaction": {
            "type": "object",
            "required": [
                "amount",
                "status"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "sellerID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tokenID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.User": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "password",
                "username"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Event"
                    }
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "ticketTransaction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.TicketTransaction"
                    }
                },
                "tokenTransaction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.TokenTransaction"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                },
                "walletTransaction": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.WalletTransaction"
                    }
                }
            }
        },
        "entities.Wallet": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "guestID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Token"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "walletType": {
                    "$ref": "#/definitions/entities.WalletType"
                },
                "walletTypeID": {
                    "type": "integer"
                }
            }
        },
        "entities.WalletPricing": {
            "type": "object",
            "required": [
                "quantity",
                "type",
                "unit_price"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "walletTypeID": {
                    "type": "integer"
                }
            }
        },
        "entities.WalletTransaction": {
            "type": "object",
            "required": [
                "quantity",
                "type"
            ],
            "properties": {
                "cashierID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "walletPricing": {
                    "$ref": "#/definitions/entities.WalletPricing"
                },
                "walletPricingID": {
                    "type": "integer"
                }
            }
        },
        "entities.WalletType": {
            "type": "object",
            "required": [
                "end_validity_date",
                "max_amount",
                "name",
                "start_validity_date"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "end_validity_date": {
                    "type": "string"
                },
                "eventID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "max_amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "online_reload": {
                    "type": "boolean"
                },
                "start_validity_date": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "walletPricings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.WalletPricing"
                    }
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}